Initial new profiling work (#74)
* Initial new profiling work

This commit introduces a new performance test `file_to_blackhole`. What we have
here is not ready for prime time yet, but the basic stubs of a repeatable
process are in place. We have indications that users are struggling with file
source capped topologies keeping up, which is why I've dropped into blackhole
versus into an http sink. As much as possible this work focuses on file source
alone.

Throughput falls off pretty rapidly as line sizes increase, looks like. See
`notes.md` in this commit for details.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Expand notes further.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Add new case

This case intends to demonstrate the difference between a multi-hop
remap config and a single hop. The remap program in vector-nohop has
problems but the basic idea is here.

Once the nohop is corrected we should also generate a differential flamegraph.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Fix config

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Update notes, capture ongoing work

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* latest updates

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>